requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.7.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-elevenlabs>=0.6.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
